Intermediate dumping, not compilable.
Min/Max moved to idMath

+++ neo/d3xp/script/Script_Program.cpp
@@ -2086,7 +2086,7 @@ void idProgram::CompileStats()
 		gameLocal.DPrintf( "   %s\n", fileList[ i ].c_str() );
 		stringspace += fileList[ i ].Allocated();
 	}
-	stringspace += fileList.Size();
+	stringspace += (int)fileList.Size();
 	
 	numdefs = varDefs.Num();
 	memused = varDefs.Num() * sizeof( idVarDef );
@@ -2095,20 +2095,20 @@ void idProgram::CompileStats()
 	
 	for( i = 0; i < types.Num(); i++ )
 	{
-		memused += types[ i ]->Allocated();
+		memused += (int)types[ i ]->Allocated();
 	}
 	
 	funcMem = functions.MemoryUsed();
 	for( i = 0; i < functions.Num(); i++ )
 	{
-		funcMem += functions[ i ].Allocated();
+		funcMem += (int)functions[ i ].Allocated();
 	}
 	
 	memallocated = funcMem + memused + sizeof( idProgram );
 	
 	memused += statements.MemoryUsed();
 	memused += functions.MemoryUsed();	// name and filename of functions are shared, so no need to include them
-	memused += sizeof( variables );
+	memused += (int)sizeof( variables );
 	
 	gameLocal.Printf( "\nMemory usage:\n" );
 	gameLocal.Printf( "     Strings: %d, %d bytes\n", fileList.Num(), stringspace );
